Fix reverse if logic when generating overlay image
Because python types are weird (and maybe because the code is actually
written poorly) we have to do "if not overlay is None:" instead of just
reversing the branches and getting rid of the "not" in the front.
Otherwise we get:

Traceback (most recent call last):
  File "plalign.py", line 68, in <module>
    if overlay:
ValueError: The truth value of an array with more than one element is ambiguous. Use a.any() or a.all()